fix!: stop generating json functions
It’s better for maintainability to avoid code generation where possible. By abstracting parts of the jsonAgg and toJsonObject functions, it’s easy enough to duplicate the other parts in dialect-specific function modules.

BREAKING CHANGE: Rename `toJsonObject` to align with dialect-specific naming. That means `jsonBuildObject` for Postgres and `jsonObject` for MySQL/SQLite.